+++ src/app/admin-panel/_actions/delete-role.tsx
@@ -1,5 +1,6 @@
 "use server";
 
+import { getBuildingName } from "@/auth/getBuildingName";
 import { getRole } from "@/auth/getRole";
 import { getToken } from "@/auth/getToken";
 import { HOST } from "@/server-endpoints/host";
@@ -8,18 +9,26 @@ import { redirect } from "next/navigation";
 export async function deleteRole(id: string) {
   const token = await getToken();
   const role = await getRole();
+  const buildingName = await getBuildingName();
+
+  if (!buildingName) {
+    throw new Error("Building name not found");
+  }
 
   if (!token || role !== "ADMIN") {
     console.error("Not authenticated");
     redirect("/login");
   }
 
-  const response = await fetch(`${HOST}/buildings/D17/roles/${id}`, {
-    method: "DELETE",
-    headers: {
-      Authorization: `Bearer ${token}`,
+  const response = await fetch(
+    `${HOST}/buildings/${buildingName}/roles/${id}`,
+    {
+      method: "DELETE",
+      headers: {
+        Authorization: `Bearer ${token}`,
+      },
     },
-  });
+  );
 
   if (!response.ok) {
     throw new Error("Failed to delete role");
